feat: create socket file
