+++ src/index.ts
@@ -4,23 +4,16 @@ import databaseConnection from "./database/connection";
 import router from "./routes/routes";
 import cors from "cors";
 import { createServer } from "http";
-import { Server } from "socket.io";
+import initializeSocket from "./sockets/socket";
 
 dotenv.config();
-
 databaseConnection();
 
 const app: Express = express();
 const port = process.env.PORT || 3000;
 
 const httpServer = createServer(app);
-
-const io = new Server(httpServer, {
-  cors: {
-    origin: process.env.ORIGIN,
-    methods: ["GET", "POST"],
-  },
-});
+initializeSocket(httpServer);
 
 app.use(
   cors({
@@ -35,20 +28,6 @@ app.use("/", router);
 
 app.get("/", (req: Request, res: Response) => {
   res.send("Express + TypeScript Server with WebSocket");
-});
-
-io.on("connection", (socket) => {
-  socket.on("sendMessage", (payload) => {
-    const { from, to } = payload;
-    let fromId = from._id;
-    let toId = to._id;
-    const roomName = [fromId, toId].sort().join("-");
-    io.to(roomName).emit("newMessage", payload);
-  });
-
-  socket.on("joinRoom", (roomName) => {
-    socket.join(roomName);
-  });
 });
 
 httpServer.listen(port, () => {
